requirements.txt
numpy
torch
matplotlib
pyserial
torchvision